Created filter map with google maps rectangle on the map to show filter area
Created algorithm to calculate first and second touch corners of rectangle for the filter option. Map clears rectangle if filter button is clicked. Once the two touches have taken place on the map after selecting filter map button, a rectangle is placed on the map. Created bounds for the filter map option
#story[1135]